use swr to select state
